requirements.txt
python-telegram-bot==20.7
feedparser==6.0.11
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.1
sqlite3-python==1.0.0
pytz==2024.1
beautifulsoup4==4.12.3
lxml==5.1.0
